chore: Header Label refactoring
