+++ packages/article/article-header/article-header-standfirst.js
@@ -20,4 +20,8 @@ HeaderStandfirst.propTypes = {
   standfirst: PropTypes.string
 };
 
+HeaderStandfirst.defaultProps = {
+  standfirst: null
+}
+
 export default HeaderStandfirst;
